improv: wrap onboarding steps in form elements and make them keyboard accessible
Previously, the onboarding was not keyboard accessible - Pressing enter
did not move to the next step. This commit wraps the onboarding steps in
a form component and changes the button to a submit so that the whole
flow is `enter`-accessible. Also to prevent default form submit
behaviour, we also need to add e.preventDefault() on the button clicks

Fixes #267