improv: wrap onboarding steps in form elements and make them keyboard accessible (#282)
* improv: wrap onboarding steps in form elements and make them keyboard accessible

Previously, the onboarding was not keyboard accessible - Pressing enter
did not move to the next step. This commit wraps the onboarding steps in
a form component and changes the button to a submit so that the whole
flow is `enter`-accessible. Also to prevent default form submit
behaviour, we also need to add e.preventDefault() on the button clicks

Fixes #267

* Update packages/core/src/components/Onboarding/pages/ExportSettings.tsx

* Update packages/core/src/components/Onboarding/pages/ImportInstructions.tsx

* Update packages/core/src/components/Onboarding/pages/NamePrimary.tsx

* Update packages/core/src/components/Onboarding/pages/OtherColors.tsx

* Update packages/core/src/components/Onboarding/pages/PickPrimary.tsx

* Update packages/core/src/components/Onboarding/pages/Referral.tsx

* Update packages/core/src/components/Onboarding/pages/ReviewPrimary.tsx

---------

Co-authored-by: Teddy Ni <teddyni19@gmail.com>